+++ src/main/java/com/nix/controller/CategoryController.java
@@ -49,7 +49,7 @@ public class CategoryController {
 	}
 
 	@GetMapping("/books/{bookId}/categories")
-	public ResponseEntity<?> getCategoriesByBook(@PathVariable Integer bookId) throws Exception {
+	public ResponseEntity<?> getCategoriesByBook(@PathVariable Long bookId) throws Exception {
 		try {
 			List<Category> categories = categoryService.findALlCategoriesByBookId(bookId);
