+++ src/main/java/com/nix/repository/RatingRepository.java
@@ -8,15 +8,15 @@ import org.springframework.data.repository.query.Param;
 
 import com.nix.models.Rating;
 
-public interface RatingRepository extends JpaRepository<Rating, Integer> {
+public interface RatingRepository extends JpaRepository<Rating, Long> {
 	
-	public List<Rating> findByBookId(Integer bookId);
+	public List<Rating> findByBookId(Long bookId);
 	
-	public List<Rating> findByUserId(Integer userId);
+	public List<Rating> findByUserId(Long userId);
 	
 	@Query("SELECT AVG(r.rating) FROM Rating r WHERE r.book.id = :bookId")
-	public Double getAverageRatingByBookId(@Param(value = "bookId") Integer bookId);
+	public Double getAverageRatingByBookId(@Param(value = "bookId") Long bookId);
 	
 	@Query("select r from Rating r where r.book.id=:bookId and r.user.id=:userId")
-	public Rating findRatingByBookAndUserId(@Param(value="bookId")Integer bookId, @Param(value="userId")Integer userId);
+	public Rating findRatingByBookAndUserId(@Param(value="bookId")Long bookId, @Param(value="userId")Long userId);
 }
