+++ src/main/java/com/nix/repository/UserFollowRepository.java
@@ -11,7 +11,7 @@ import com.nix.models.UserFollow;
 public interface UserFollowRepository extends JpaRepository<UserFollow, Long> {
 	Optional<UserFollow> findByFollowerAndFollowed(User follower, User followed);
 
-	List<UserFollow> findByFollowedId(Integer userId);
+	List<UserFollow> findByFollowedId(Long userId);
 
-	List<UserFollow> findByFollowerId(Integer userId);
+	List<UserFollow> findByFollowerId(Long userId);
 }
